Moved Files to hidden folders.
This makes the Include Library only include the two main headers
