+++ library.properties
@@ -1,5 +1,5 @@
 name=NeoPixelBus by Makuna
-version=2.0.1
+version=2.0.2
 author=Michael C. Miller (makuna@live.com)
 maintainer=Michael C. Miller (makuna@live.com)
 sentence=A library that makes controlling NeoPixels (WS2811, WS2812 & SK6812) easy.
