+++ src/NeoPixelAnimator.h
@@ -27,7 +27,7 @@ License along with NeoPixel.  If not, see
 #pragma once
 
 #include <Arduino.h>
-#include "NeoEase.h"
+#include "internal/NeoEase.h"
 
 enum AnimationState
 {
