+++ src/internal/HsbColor.cpp
