+++ src/internal/HsbColor.h
