+++ src/internal/NeoPixelAnimator.cpp
