+++ src/internal/RgbColor.cpp
