+++ src/internal/RgbColor.h
