+++ src/internal/RgbwColor.cpp
