+++ src/internal/RgbwColor.h
